try to fix uwp code projects
