+++ samples/snippets/csharp/VS_Snippets_CLR_System/WindowsRuntimeStreamExtensionsEx/MainPage.xaml
@@ -1,0 +1,17 @@
+﻿<Page
+    x:Class="UWPApp2.MainPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:UWPApp2"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+
+    <Grid>
+        <TextBlock x:Name="TextBlock2" HorizontalAlignment="Left" Margin="184,124,0,0" Text="TextBlock" TextWrapping="Wrap" VerticalAlignment="Top"/>
+        <Button x:Name="Button2" Click="button2_Click" Content="Button" Margin="284,342,0,0" VerticalAlignment="Top"/>
+        <Image x:Name="Image1" HorizontalAlignment="Left" Height="100" Margin="538,252,0,0" VerticalAlignment="Top" Width="100"/>
+
+    </Grid>
+</Page>
