+++ samples/snippets/csharp/VS_Snippets_CLR_System/WindowsRuntimeStreamExtensionsEx/snippets5000.json
@@ -1,0 +1,3 @@
+{
+    "host": "visualstudio"
+}
