+++ samples/snippets/csharp/VS_Snippets_CLR_System/system.io.windowsruntimestreamextensionsex/App.xaml
@@ -1,0 +1,7 @@
+﻿<Application
+    x:Class="UWPApp.App"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:UWPApp">
+
+</Application>
