+++ samples/snippets/csharp/VS_Snippets_CLR_System/system.io.windowsruntimestreamextensionsex/MainPage.xaml
@@ -1,0 +1,16 @@
+﻿<Page
+    x:Class="UWPApp.MainPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:UWPApp"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+
+    <Grid>
+        <TextBlock x:Name="TextBlock1" HorizontalAlignment="Left" Margin="-1988,15,0,0" Text="TextBlock" TextWrapping="Wrap" VerticalAlignment="Top"/>
+        <Button x:Name="Button1" Click="button1_Click" Content="Button" Margin="1800,500,0,0" VerticalAlignment="Top" Width="0"/>
+
+    </Grid>
+</Page>
