--- samples/snippets/csharp/VS_Snippets_CLR_System/system.io.windowsruntimestreamextensionsex/cs/app.xaml
@@ -1,7 +1,0 @@
-﻿<Application
-    x:Class="TextFile.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:TextFile">
-
-</Application>
