--- samples/snippets/csharp/VS_Snippets_CLR_System/system.io.windowsruntimestreamextensionsex/cs/package.appxmanifest
@@ -1,44 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Package xmlns="http://schemas.microsoft.com/appx/2010/manifest" xmlns:m2="http://schemas.microsoft.com/appx/2013/manifest">
-  <Identity Name="e6404d5b-6f05-40d0-8411-d328f6e2d84f" Publisher="CN=cherylsi" Version="1.0.0.0" />
-  <Properties>
-    <DisplayName>TextFile</DisplayName>
-    <PublisherDisplayName>cherylsi</PublisherDisplayName>
-    <Logo>Assets\StoreLogo.png</Logo>
-  </Properties>
-  <Prerequisites>
-    <OSMinVersion>6.3.0</OSMinVersion>
-    <OSMaxVersionTested>6.3.0</OSMaxVersionTested>
-  </Prerequisites>
-  <Resources>
-    <Resource Language="x-generate" />
-  </Resources>
-  <Applications>
-    <Application Id="App" Executable="$targetnametoken$.exe" EntryPoint="TextFile.App">
-      <m2:VisualElements DisplayName="TextFile" 
-                         Square150x150Logo="Assets\Logo.png" 
-                         Square30x30Logo="Assets\SmallLogo.png" 
-                         Description="TextFile" 
-                         ForegroundText="light" 
-                         BackgroundColor="#464646">
-        <m2:SplashScreen Image="Assets\SplashScreen.png" />
-      </m2:VisualElements>
-      <!--<snippetManifest>-->
-      <Extensions>
-        <Extension Category="windows.fileTypeAssociation">
-          <FileTypeAssociation Name="txt">
-            <DisplayName>Text</DisplayName>
-            <SupportedFileTypes>
-              <FileType>.txt</FileType>
-            </SupportedFileTypes>
-          </FileTypeAssociation>
-        </Extension>
-      </Extensions>
-    </Application>
-  </Applications>
-  <Capabilities>
-    <Capability Name="internetClient" />
-    <Capability Name="documentsLibrary" />
-  </Capabilities>
-  <!--</snippetManifest>-->
-</Package>
