--- samples/snippets/visualbasic/VS_Snippets_CLR_System/system.io.windowsruntimestreamextensionsex/vb/app.xaml
@@ -1,7 +1,0 @@
-﻿<Application
-    x:Class="TextFileVB.App"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:TextFileVB">
-
-</Application>
