--- samples/snippets/visualbasic/VS_Snippets_CLR_System/system.io.windowsruntimestreamextensionsex/vb/mainpage.xaml
@@ -1,17 +1,0 @@
-﻿<Page
-    x:Class="TextFileVB.MainPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:TextFileVB"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d">
-
-    <Grid Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
-        <TextBlock Name="TextBlock1" HorizontalAlignment="Left" Margin="106,210,0,0" TextWrapping="Wrap"  VerticalAlignment="Top" Height="362" Width="914"/>
-        <Button x:Name="button1" Content=".NET to WinRT" HorizontalAlignment="Left" Margin="509,237,0,0" VerticalAlignment="Top" Click="button1_Click" Height="94" Width="243"/>
-        <Button x:Name="button2" Click="button2_Click"  Content="WinRT to .NET" HorizontalAlignment="Left" Margin="512,330,0,0" VerticalAlignment="Top" Height="142" Width="240"/>
-        <Image Name="image1" HorizontalAlignment="Left" Height="230" Margin="73,240,0,0" VerticalAlignment="Top" Width="400"/>
-
-    </Grid>
-</Page>
