--- samples/snippets/visualbasic/VS_Snippets_CLR_System/system.io.windowsruntimestreamextensionsex/vb/package.appxmanifest
@@ -1,39 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Package xmlns="http://schemas.microsoft.com/appx/2010/manifest" xmlns:m2="http://schemas.microsoft.com/appx/2013/manifest">
-  <Identity Name="0ba44a27-e500-4d1d-8862-3ab65f32728b" Publisher="CN=cherylsi" Version="1.0.0.0" />
-  <Properties>
-    <DisplayName>TextFileVB</DisplayName>
-    <PublisherDisplayName>cherylsi</PublisherDisplayName>
-    <Logo>Assets\StoreLogo.png</Logo>
-  </Properties>
-  <Prerequisites>
-    <OSMinVersion>6.3.0</OSMinVersion>
-    <OSMaxVersionTested>6.3.0</OSMaxVersionTested>
-  </Prerequisites>
-  <Resources>
-    <Resource Language="x-generate" />
-  </Resources>
-  <Applications>
-    <Application Id="App" Executable="$targetnametoken$.exe" EntryPoint="TextFileVB.App">
-      <m2:VisualElements DisplayName="TextFileVB" Square150x150Logo="Assets\Logo.png" Square30x30Logo="Assets\SmallLogo.png" Description="TextFileVB" ForegroundText="light" BackgroundColor="#464646">
-        <m2:SplashScreen Image="Assets\SplashScreen.png" />
-      </m2:VisualElements>
-      <Extensions>
-        <Extension Category="windows.fileTypeAssociation">
-          <FileTypeAssociation Name="txt">
-            <DisplayName>Text</DisplayName>
-            <SupportedFileTypes>
-              <FileType>.txt</FileType>
-            </SupportedFileTypes>
-          </FileTypeAssociation>
-        </Extension>
-      </Extensions>
-    </Application>
-  </Applications>
-  <Capabilities>
-    <Capability Name="internetClient" />
-    <Capability Name="picturesLibrary" />
-    <Capability Name="documentsLibrary" />
-    <Capability Name="musicLibrary" />
-  </Capabilities>
-</Package>
